+++ code/src/nuvla/connector/__init__.py
@@ -1,15 +1,19 @@
 # -*- coding: utf-8 -*-
 
+from typing import Optional
+from nuvla.api.api import Api as Nuvla
 
-def connector_factory(connector_class, nuvla_api, credential_id, api_infrastructure_service=None):
+
+def connector_factory(connector, nuvla_api: Nuvla, credential_id,
+                      infrastructure_service: Optional[dict]=None):
     """
     Returns infrastructure connector instance. Connector to be used is derived
     from Nuvla via `credential_id`.
 
-    :param connector_class: Connector class to instanciate with
+    :param connector: Connector (module) to instantiate with
     :param nuvla_api: nuvla.api.Api, instance
     :param credential_id: str
-    :param api_infrastructure_service: if available, the corresponding infrastructure service
+    :param infrastructure_service: if available, the corresponding infrastructure service
     :return: connector.Connector, instance
     """
 
@@ -19,11 +23,11 @@ def connector_factory(connector_class, nuvla_api, credential_id, api_infrastruct
     if credential_id is None:
         raise ValueError("Credential id is not provided!")
 
-    api_credential = nuvla_api.get(credential_id).data
+    credential = nuvla_api.get(credential_id).data
 
-    if not api_infrastructure_service:
-        infrastructure_service_id = api_credential['parent']
+    if not infrastructure_service:
+        infrastructure_service_id = credential['parent']
 
-        api_infrastructure_service = nuvla_api.get(infrastructure_service_id).data
+        infrastructure_service = nuvla_api.get(infrastructure_service_id).data
 
-    return connector_class.instantiate_from_cimi(api_infrastructure_service, api_credential)
+    return connector.instantiate_from_cimi(infrastructure_service, credential)